refactor: separating widgets
