+++ frontend/lib/pages/startup_details.dart
@@ -1,16 +1,18 @@
 //Autor: Pedro Vinicius Romanato & Allan Giovanni Matias Paes
 
-import 'dart:ui' as ui;
 import 'package:flutter/material.dart';
 import 'package:frontend/widgets/feedback_modal.dart';
+import 'package:frontend/widgets/socio_row.dart';
 import 'package:intl/intl.dart';
 import '../models/startup.dart';
 import '../services/startup_service.dart';
 import '../widgets/price_chart.dart';
+import '../widgets/startup_fixed_header.dart';
+import '../widgets/card_sobre_startup.dart';
+import '../widgets/demo_video_player.dart';
+import '../widgets/documents_download_card.dart';
 import './faq_page.dart';
 import 'package:frontend/pages/buy_from_startup_page.dart';
-import 'package:video_player/video_player.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 class StartupDetailsPage extends StatefulWidget {
   final String startupId;
@@ -396,12 +398,11 @@ class _StartupDetailsPageState extends State<StartupDetailsPage> {
                       const SizedBox(height: 15),
                       ...data.founders.map((founder) => Padding(
                             padding: const EdgeInsets.only(bottom: 15),
-                            child: _buildSocioRow(
-                              context,
-                              gerarIniciais(founder.name),
-                              founder.name,
-                              founder.role,
-                              "${founder.equityPercent}%",
+                            child: SocioRow(
+                              iniciais: gerarIniciais(founder.name),
+                              nome: founder.name,
+                              cargo: founder.role,
+                              porcentagem: "${founder.equityPercent}%",
                             ),
                           )),
                     ],
@@ -498,537 +499,6 @@ class _StartupDetailsPageState extends State<StartupDetailsPage> {
                 fontSize: 17,
                 color: theme.colorScheme.onSurface)),
       ],
-    );
-  }
-}
-
-class FixedHeader extends StatelessWidget implements PreferredSizeWidget {
-  final String name;
-  final String segment;
-  final String logoPath;
-  final String shortDescription;
-
-  const FixedHeader(
-      {super.key,
-      required this.name,
-      required this.segment,
-      required this.logoPath,
-      required this.shortDescription});
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Container(
-      padding: EdgeInsets.only(
-          top: MediaQuery.of(context).padding.top + 10,
-          left: 20,
-          right: 20,
-          bottom: 15),
-      decoration: BoxDecoration(
-          color: theme.colorScheme.surface,
-          border: Border(bottom: BorderSide(color: theme.dividerColor.withOpacity(0.1)))),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          IconButton(
-            icon: Icon(Icons.arrow_back, color: theme.colorScheme.onSurface),
-            onPressed: () => Navigator.maybePop(context),
-            padding: EdgeInsets.zero,
-            constraints: const BoxConstraints(),
-          ),
-          const SizedBox(height: 15),
-          Row(
-            children: [
-              ClipRRect(
-                borderRadius: BorderRadius.circular(12),
-                child: Image.network(
-                  logoPath,
-                  height: 50,
-                  width: 50,
-                  fit: BoxFit.cover,
-                  errorBuilder: (_, __, ___) =>
-                      Icon(Icons.business, size: 50, color: theme.colorScheme.onSurfaceVariant),
-                ),
-              ),
-              const SizedBox(width: 15),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(name,
-                        style: TextStyle(
-                            fontWeight: FontWeight.bold, fontSize: 18, color: theme.colorScheme.onSurface)),
-                    const SizedBox(height: 4),
-                    Container(
-                      padding: const EdgeInsets.symmetric(
-                          horizontal: 8, vertical: 2),
-                      decoration: BoxDecoration(
-                        color: const Color(0xFF00A84E).withOpacity(0.1),
-                        borderRadius: BorderRadius.circular(6),
-                      ),
-                      child: Text(
-                        segment.replaceFirst(segment[0], segment[0].toUpperCase()),
-                        style: const TextStyle(
-                          color: Color(0xFF00A84E),
-                          fontWeight: FontWeight.bold,
-                          fontSize: 13,
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 12),
-          Text(
-            shortDescription,
-            maxLines: 2,
-            overflow: TextOverflow.ellipsis,
-            style: TextStyle(
-                fontWeight: FontWeight.bold, fontSize: 13, color: theme.colorScheme.onSurfaceVariant),
-          ),
-        ],
-      ),
-    );
-  }
-
-  @override
-  Size get preferredSize => const Size.fromHeight(200);
-}
-
-class CardSobreStartup extends StatefulWidget {
-  final String descricao;
-  const CardSobreStartup({super.key, required this.descricao});
-
-  @override
-  State<CardSobreStartup> createState() => _CardSobreStartupState();
-}
-
-class _CardSobreStartupState extends State<CardSobreStartup> {
-  bool isExpanded = false;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Container(
-      padding: const EdgeInsets.all(15),
-      decoration:
-          BoxDecoration(color: theme.colorScheme.surface, border: Border.all(color: theme.dividerColor.withOpacity(0.1)), borderRadius: BorderRadius.circular(15)),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text('Sobre a startup',
-              style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18, color: theme.colorScheme.onSurface)),
-          const SizedBox(height: 8),
-          LayoutBuilder(
-            builder: (context, constraints) {
-              final span = TextSpan(
-                text: widget.descricao,
-                style: TextStyle(color: theme.colorScheme.onSurfaceVariant, height: 1.5),
-              );
-              final tp = TextPainter(
-                text: span,
-                maxLines: 3,
-                textAlign: TextAlign.left,
-                textDirection: ui.TextDirection.ltr,
-              );
-              tp.layout(maxWidth: constraints.maxWidth);
-              final needsExpansion = tp.didExceedMaxLines;
-
-              return Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    widget.descricao,
-                    maxLines: isExpanded ? null : 3,
-                    overflow:
-                        isExpanded ? TextOverflow.visible : TextOverflow.ellipsis,
-                    style: TextStyle(color: theme.colorScheme.onSurfaceVariant, height: 1.5),
-                  ),
-                  if (needsExpansion)
-                    Padding(
-                      padding: const EdgeInsets.only(top: 8.0),
-                      child: InkWell(
-                        onTap: () => setState(() => isExpanded = !isExpanded),
-                        child: Row(
-                          children: [
-                            Text(isExpanded ? 'Ver menos' : 'Ver mais',
-                                style: const TextStyle(
-                                    color: Color(0xFF00A84E),
-                                    fontWeight: FontWeight.bold)),
-                            Icon(
-                                isExpanded
-                                    ? Icons.keyboard_arrow_up
-                                    : Icons.keyboard_arrow_down,
-                                color: const Color(0xFF00A84E)),
-                          ],
-                        ),
-                      ),
-                    ),
-                ],
-              );
-            },
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-Widget _buildSocioRow(
-    BuildContext context, String iniciais, String nome, String cargo, String porcentagem) {
-  final theme = Theme.of(context);
-  return Row(
-    children: [
-      CircleAvatar(
-        radius: 22,
-        backgroundColor: const Color(0xFF00A84E).withOpacity(0.1),
-        child: Text(iniciais,
-            style: const TextStyle(
-                color: Color(0xFF00A84E), fontWeight: FontWeight.bold)),
-      ),
-      const SizedBox(width: 12),
-      Expanded(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(nome,
-                style:
-                    TextStyle(fontWeight: FontWeight.bold, fontSize: 16, color: theme.colorScheme.onSurface)),
-            Text(cargo,
-                style: TextStyle(fontSize: 13, color: theme.colorScheme.onSurfaceVariant)),
-          ],
-        ),
-      ),
-      Text(porcentagem,
-          style: const TextStyle(
-              color: Color(0xFF00A84E),
-              fontWeight: FontWeight.bold,
-              fontSize: 16)),
-    ],
-  );
-}
-class DemoVideoPlayer extends StatefulWidget {
-  final String videoUrl;
-  const DemoVideoPlayer({super.key, required this.videoUrl});
-
-  @override
-  State<DemoVideoPlayer> createState() => _DemoVideoPlayerState();
-}
-
-class _DemoVideoPlayerState extends State<DemoVideoPlayer> {
-  late VideoPlayerController _controller;
-  bool _hasError = false;
-  bool _isInitialized = false;
-
-  @override
-  void initState() {
-    super.initState();
-
-    if (!widget.videoUrl.startsWith('https://firebasestorage.googleapis.com/')) {
-      setState(() => _hasError = true);
-      return;
-    }
-
-    _controller = VideoPlayerController.networkUrl(Uri.parse(widget.videoUrl))
-      ..initialize().then((_) {
-        if (mounted) {
-          setState(() {
-            _isInitialized = true;
-          });
-        }
-      }).catchError((error) {
-        if (mounted) {
-          setState(() => _hasError = true);
-        }
-      });
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    if (_hasError) {
-      return SizedBox(
-        height: 200,
-        child: Center(
-          child: Text(
-            "Vídeo Indisponível",
-            textAlign: TextAlign.center,
-            style: TextStyle(color: theme.colorScheme.error),
-          ),
-        ),
-      );
-    }
-
-    if (!_isInitialized) {
-      return const SizedBox(
-        height: 200,
-        child: Center(child: CircularProgressIndicator(color: Color(0xFF00A84E))),
-      );
-    }
-
-    return Column(
-      children: [
-        ClipRRect(
-          borderRadius: BorderRadius.circular(12),
-          child: AspectRatio(
-            aspectRatio: _controller.value.aspectRatio,
-            child: Stack(
-              alignment: Alignment.bottomCenter,
-              children: [
-                VideoPlayer(_controller),
-                _ControlsOverlay(controller: _controller),
-                VideoProgressIndicator(
-                  _controller,
-                  allowScrubbing: true,
-                  colors: const VideoProgressColors(
-                    playedColor: Color(0xFF00A84E),
-                    bufferedColor: Colors.white24,
-                    backgroundColor: Colors.white12,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _ControlsOverlay extends StatefulWidget {
-  const _ControlsOverlay({required this.controller});
-
-  final VideoPlayerController controller;
-
-  @override
-  State<_ControlsOverlay> createState() => _ControlsOverlayState();
-}
-
-class _ControlsOverlayState extends State<_ControlsOverlay> {
-  @override
-  void initState() {
-    super.initState();
-    widget.controller.addListener(_updateOverlay);
-  }
-
-  @override
-  void didUpdateWidget(_ControlsOverlay oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.controller != widget.controller) {
-      oldWidget.controller.removeListener(_updateOverlay);
-      widget.controller.addListener(_updateOverlay);
-    }
-  }
-
-  @override
-  void dispose() {
-    widget.controller.removeListener(_updateOverlay);
-    super.dispose();
-  }
-
-  void _updateOverlay() {
-    if (mounted) setState(() {});
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final bool isFinished = widget.controller.value.isInitialized &&
-        widget.controller.value.position >= widget.controller.value.duration;
-
-    return GestureDetector(
-      behavior: HitTestBehavior.opaque,
-      onTap: () {
-        if (isFinished) {
-          widget.controller.seekTo(Duration.zero);
-          widget.controller.play();
-        } else {
-          widget.controller.value.isPlaying
-              ? widget.controller.pause()
-              : widget.controller.play();
-        }
-      },
-      child: Stack(
-        children: <Widget>[
-          AnimatedSwitcher(
-            duration: const Duration(milliseconds: 50),
-            reverseDuration: const Duration(milliseconds: 200),
-            child: widget.controller.value.isPlaying
-                ? const SizedBox.shrink()
-                : Container(
-                    color: Colors.black26,
-                    child: Center(
-                      child: Icon(
-                        isFinished ? Icons.replay : Icons.play_arrow,
-                        color: Colors.white,
-                        size: 100.0,
-                        semanticLabel: isFinished ? 'Replay' : 'Play',
-                      ),
-                    ),
-                  ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-
-class DocumentsDownloadCard extends StatelessWidget {
-  final String? pitchDeckUrl;
-  final String? businessPlanUrl;
-  final String? executiveSummary;
-
-  const DocumentsDownloadCard({
-    super.key,
-    this.pitchDeckUrl,
-    this.businessPlanUrl,
-    this.executiveSummary,
-  });
-
-  bool get _hasAny =>
-      pitchDeckUrl != null ||
-      businessPlanUrl != null ||
-      executiveSummary != null;
-
-  Future<void> _launch(BuildContext context, String? url, String label) async {
-    if (url == null || url.isEmpty) {
-      FeedbackModal.show(
-        context: context,
-        title: 'Indisponível',
-        message: '$label não disponível',
-        type: FeedbackType.info,
-      );
-      return;
-    }
-    final uri = Uri.parse(url);
-    if (!await launchUrl(uri, mode: LaunchMode.externalApplication)) {
-      if (context.mounted) {
-        FeedbackModal.show(
-          context: context,
-          title: 'Erro',
-          message: 'Não foi possível abrir $label',
-          type: FeedbackType.error,
-        );
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (!_hasAny) return const SizedBox.shrink();
-    final theme = Theme.of(context);
-
-    return Container(
-      padding: const EdgeInsets.all(15),
-      decoration: BoxDecoration(
-        color: theme.colorScheme.surface,
-        border: Border.all(color: theme.dividerColor.withOpacity(0.1)),
-        borderRadius: BorderRadius.circular(15),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            'Documentos',
-            style: TextStyle(
-                fontWeight: FontWeight.bold,
-                fontSize: 18,
-                color: theme.colorScheme.onSurface),
-          ),
-          const SizedBox(height: 12),
-          _DocItem(
-            icon: Icons.slideshow_rounded,
-            label: 'Apresentação para investidores',
-            available: pitchDeckUrl != null,
-            onTap: () => _launch(context, pitchDeckUrl, 'Pitch Deck'),
-          ),
-          Divider(height: 20, color: theme.dividerColor.withOpacity(0.1)),
-          _DocItem(
-            icon: Icons.description_rounded,
-            label: 'Plano de negócios',
-            available: businessPlanUrl != null,
-            onTap: () => _launch(context, businessPlanUrl, 'Business Plan'),
-          ),
-          Divider(height: 20, color: theme.dividerColor.withOpacity(0.1)),
-          _DocItem(
-            icon: Icons.summarize_rounded,
-            label: 'Resumo Executivo',
-            available: executiveSummary != null,
-            onTap: () =>
-                _launch(context, executiveSummary, 'Executive Summary'),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _DocItem extends StatelessWidget {
-  final IconData icon;
-  final String label;
-  final bool available;
-  final VoidCallback onTap;
-
-  const _DocItem({
-    required this.icon,
-    required this.label,
-    required this.available,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final color = available ? const Color(0xFF00A84E) : theme.colorScheme.onSurfaceVariant.withOpacity(0.5);
-
-    return InkWell(
-      onTap: available ? onTap : null,
-      borderRadius: BorderRadius.circular(8),
-      child: Padding(
-        padding: const EdgeInsets.symmetric(vertical: 4),
-        child: Row(
-          children: [
-            Container(
-              padding: const EdgeInsets.all(8),
-              decoration: BoxDecoration(
-                color: available
-                    ? const Color(0xFF00A84E).withOpacity(0.1)
-                    : theme.colorScheme.surfaceVariant.withOpacity(0.3),
-                borderRadius: BorderRadius.circular(8),
-              ),
-              child: Icon(icon, color: color, size: 22),
-            ),
-            const SizedBox(width: 12),
-            Expanded(
-              child: Text(
-                label,
-                style: TextStyle(
-                  fontWeight: FontWeight.w600,
-                  fontSize: 15,
-                  color: available ? theme.colorScheme.onSurface : theme.colorScheme.onSurfaceVariant.withOpacity(0.5),
-                ),
-              ),
-            ),
-            Icon(
-              available
-                  ? Icons.download_rounded
-                  : Icons.lock_outline_rounded,
-              color: color,
-              size: 20,
-            ),
-          ],
-        ),
-      ),
     );
   }
 }
