+++ frontend/lib/widgets/startup_fixed_header.dart
@@ -1,0 +1,105 @@
+import 'package:flutter/material.dart';
+
+class FixedHeader extends StatelessWidget implements PreferredSizeWidget {
+  final String name;
+  final String segment;
+  final String logoPath;
+  final String shortDescription;
+
+  const FixedHeader({
+    super.key,
+    required this.name,
+    required this.segment,
+    required this.logoPath,
+    required this.shortDescription,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Container(
+      padding: EdgeInsets.only(
+          top: MediaQuery.of(context).padding.top + 10,
+          left: 20,
+          right: 20,
+          bottom: 15),
+      decoration: BoxDecoration(
+          color: theme.colorScheme.surface,
+          border: Border(
+              bottom:
+                  BorderSide(color: theme.dividerColor.withOpacity(0.1)))),
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          IconButton(
+            icon: Icon(Icons.arrow_back, color: theme.colorScheme.onSurface),
+            onPressed: () => Navigator.maybePop(context),
+            padding: EdgeInsets.zero,
+            constraints: const BoxConstraints(),
+          ),
+          const SizedBox(height: 15),
+          Row(
+            children: [
+              ClipRRect(
+                borderRadius: BorderRadius.circular(12),
+                child: Image.network(
+                  logoPath,
+                  height: 50,
+                  width: 50,
+                  fit: BoxFit.cover,
+                  errorBuilder: (_, __, ___) => Icon(Icons.business,
+                      size: 50, color: theme.colorScheme.onSurfaceVariant),
+                ),
+              ),
+              const SizedBox(width: 15),
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(name,
+                        style: TextStyle(
+                            fontWeight: FontWeight.bold,
+                            fontSize: 18,
+                            color: theme.colorScheme.onSurface)),
+                    const SizedBox(height: 4),
+                    Container(
+                      padding: const EdgeInsets.symmetric(
+                          horizontal: 8, vertical: 2),
+                      decoration: BoxDecoration(
+                        color: const Color(0xFF00A84E).withOpacity(0.1),
+                        borderRadius: BorderRadius.circular(6),
+                      ),
+                      child: Text(
+                        segment.replaceFirst(
+                            segment[0], segment[0].toUpperCase()),
+                        style: const TextStyle(
+                          color: Color(0xFF00A84E),
+                          fontWeight: FontWeight.bold,
+                          fontSize: 13,
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+          const SizedBox(height: 12),
+          Text(
+            shortDescription,
+            maxLines: 2,
+            overflow: TextOverflow.ellipsis,
+            style: TextStyle(
+                fontWeight: FontWeight.bold,
+                fontSize: 13,
+                color: theme.colorScheme.onSurfaceVariant),
+          ),
+        ],
+      ),
+    );
+  }
+
+  @override
+  Size get preferredSize => const Size.fromHeight(200);
+}
